# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UPS e-Recruit API",
    "version": "1.0.0",
    "description": "Versioned same-origin API for applicant and authorised recruitment operations. OCR is assistive evidence only."
  },
  "servers": [{ "url": "/api" }],
  "security": [{ "sanctum": [] }],
  "paths": {
...
    "/v1/operations/lookups": { "get": { "operationId": "listOperationsLookups", "responses": { "200": { "description": "Role- and scope-filtered human-readable operational registers plus the headquarters hard-copy receipt capability and receiving point" }, "403": { "$ref": "#/components/responses/Forbidden" } } } },
...
    "/v1/selection/lookups": { "get": { "operationId": "listSelectionPolicyOptions", "responses": { "200": { "description": "Human-readable completed ranking runs, quota buckets and verified skill categories for structured selection policy entry" }, "403": { "$ref": "#/components/responses/Forbidden" } } } },
...
  },
  "components": {
    "securitySchemes": { "sanctum": { "type": "http", "scheme": "bearer" } },
...
    "responses": {
...
      "Forbidden": { "description": "Role/scope denied" },
...
    }
  }
}